relations

+++ app/Http/Controllers/ProductController.php
@@ -7,7 +7,8 @@ use App\Models\Product;
 use Illuminate\Support\Facades\DB;
 use App\Models\Cart;
 use App\Models\Order;
-use Session;
+use Illuminate\Support\Facades\Session;
+
 
 class ProductController extends Controller
 {
@@ -42,7 +43,7 @@ class ProductController extends Controller
         {
             return redirect('/login');
         }
-        
+
     }
     static function cartItem()
     {
@@ -72,36 +73,45 @@ class ProductController extends Controller
          ->join('products','cart.product_id','=','products.id')
          ->where('cart.user_id',$userId)
          ->sum('products.price');
- 
+
          return view('ordernow',['total'=>$total]);
     }
     function orderPlace(Request $req)
     {
         $userId=Session::get('user')['id'];
          $allCart= Cart::where('user_id',$userId)->get();
+         if ($req->address == null){
+
+             flash('Data submitted')->important();
+             return redirect()->back();
+         }
+         else {
          foreach($allCart as $cart)
          {
              $order= new Order;
-             $order->product_id=$cart['product_id'];
              $order->user_id=$cart['user_id'];
+
              $order->status="pending";
              $order->payment_method=$req->payment;
              $order->payment_status="pending";
              $order->address=$req->address;
              $order->save();
-             Cart::where('user_id',$userId)->delete(); 
+             $order->products()->attach($cart['product_id']);
+
+             Cart::where('user_id',$userId)->delete();
          }
          $req->input();
-         return redirect('/');
+         return redirect('/');}
     }
     function myOrders()
     {
         $userId=Session::get('user')['id'];
-        $orders= DB::table('orders')
-         ->join('products','orders.product_id','=','products.id')
-         ->where('orders.user_id',$userId)
-         ->get();
- 
+//        $orders= DB::table('orders')
+//         ->join('products','orders.product_id','=','products.id')
+//         ->where('orders.user_id',$userId)
+//         ->get();
+        $orders=  Order::where('orders.user_id',$userId)->with('products')->get();
+        dd($orders);
          return view('myorders',['orders'=>$orders]);
     }
 }
